mock set_constants_to_default method
