+++ scripts/start.py
@@ -159,7 +159,7 @@ def main(args: Optional[Sequence[str]] = None) -> None:
         if parsed_args.source_maps:
             build_args.append('--source_maps')
         build.main(args=build_args)
-        stack.callback(build.set_constants_to_default)
+        stack.callback(common.set_constants_to_default)
 
         stack.enter_context(servers.managed_redis_server())
         stack.enter_context(servers.managed_elasticsearch_dev_server())
